+++ src/example/app.component.ts
@@ -1,99 +1,38 @@
-import { Component, ViewChild } from '@angular/core'
+import { Component } from '@angular/core'
 import { exampleData } from './example-table.data'
-import { exampleTableConfig } from './example-table.config'
-import { Ng2TableComponent } from '../../'
+import { TableConfigModel } from '../../index'
 
 @Component({
   selector: 'app',
   template: `
-    <button class="btn btn-secondary btn-sm"
-      (click)="toggleTable()">
-      Toggle table
-    </button>
-    <button class="btn btn-secondary btn-sm"
-      (click)="reorganizeContent()">
-      Reorganize content and remove 2 random items
-    </button>
-    <button class="btn btn-secondary btn-sm"
-      (click)="activateRow()">
-      Activate row
-    </button>
-    <button class="btn btn-secondary btn-sm"
-      *ngIf="tableConfigUpdatedCounter > 0"
-      (click)="setTableConfigToUpdated()">
-      Set tableConfig to updated
-      {{tableConfigUpdatedCounter}}
-    </button>
-    <span>
-    </span>
+    <example-table-component>
+    </example-table-component>
 
     <br><br>
+    <button class="btn btn-secondary btn-sm" (click)="showMinimalTableFn()">
+      Show minimal config example
+    </button>
 
-    <!-- Set the height of the table content on a wrapping div -->
-    <div style="height: 340px;" *ngIf="showTable">
-      <ng2-table
-        [tableData]="tableData"
-        [tableConfig]="tableConfig"
-        (rowClicked)="rowClicked($event)"
-        (cellItemClicked)="cellItemClicked($event)"
-        (tableConfigUpdated)="tableConfigUpdated($event)">
-      </ng2-table>
-    </div>
+    <ng2-table
+      *ngIf="showMinimalTable"
+      [tableData]="exampleData"
+      [tableConfig]="minimalTableConfig">
+    </ng2-table>
   `
 })
 export class AppComponent {
-  public showTable = true
-  public tableData = exampleData
-  public tableConfig = exampleTableConfig
-  public tableConfigUpdatedCounter = 0
-  public updatedTableConfig
-  public rowToActivate
-
-  @ViewChild(Ng2TableComponent) public ng2TableComponent: Ng2TableComponent
-
-  public rowClicked ($event) {
-    console.log('(rowClicked):', $event)
-  }
-  public cellItemClicked (options) {
-    console.log('(cellItemClicked):', options)
-  }
-  public tableConfigUpdated (config) {
-    console.log('(tableConfigUpdated):', config)
-    this.tableConfigUpdatedCounter++
-    this.updatedTableConfig = config
+  public exampleData = exampleData
+  public showMinimalTable = false
+  public minimalTableConfig: TableConfigModel = {
+    columnDefs: [
+      { field: 'userId' },
+      { field: 'pet' },
+      { field: 'userName' },
+      { field: 'nickName' }
+    ]
   }
 
-  public toggleTable () {
-    this.showTable = !this.showTable
-  }
-  public setTableConfigToUpdated () {
-    this.tableConfig = this.updatedTableConfig
-  }
-  public activateRow () {
-    this.ng2TableComponent.activateRow({
-      rowIndex: this.getRandomInt(0, this.tableData.length - 1)
-    })
-  }
-
-  public reorganizeContent () {
-    let order = this.getRandomInt(0, 1)
-    let dataCopy = exampleData.map((item) => {
-      return Object.assign({}, item)
-    })
-    dataCopy.sort((item1, item2) => {
-      if (item1.userId === item2.userId) {
-        return 0
-      } else if (item1.userId > item2.userId) {
-        return order === 1 ? 1 : -1
-      } else {
-        return order === 1 ? -1 : 1
-      }
-    })
-    dataCopy.splice(this.getRandomInt(0, 4), 1)
-    dataCopy.splice(this.getRandomInt(0, 3), 1)
-    this.tableData = dataCopy
-  }
-  public getRandomInt (min, max) {
-    return Math.floor(Math.random() * (max - min + 1)) + min
+  public showMinimalTableFn () {
+    this.showMinimalTable = true
   }
 }