+++ src/example/app.module.js
@@ -11,6 +11,7 @@ var __metadata = (this && this.__metadata) || function (k, v) {
 var core_1 = require("@angular/core");
 var platform_browser_1 = require("@angular/platform-browser");
 var app_component_1 = require("./app.component");
+var example_table_component_1 = require("./example-table.component");
 var _1 = require("../../");
 var AppModule = (function () {
     function AppModule() {
@@ -19,7 +20,7 @@ var AppModule = (function () {
 }());
 AppModule = __decorate([
     core_1.NgModule({
-        declarations: [app_component_1.AppComponent],
+        declarations: [app_component_1.AppComponent, example_table_component_1.ExampleTableComponent],
         imports: [platform_browser_1.BrowserModule, _1.Ng2TableModule],
         bootstrap: [app_component_1.AppComponent]
     }),