+++ src/example/example-table.config.js
@@ -31,13 +31,14 @@ exports.exampleTableConfig = {
             field: 'userName',
             headerText: 'Name',
             filterEnabled: true,
-            filterPlaceholder: 'test',
+            filterPlaceholder: 'Filter name',
             cellNgStyle: { 'font-weight': 'bold' }
         },
         {
             field: 'nickName',
             headerText: 'Nickname',
             filterEnabled: true,
+            filterPlaceholder: 'Filter nickname',
             cellItem: {
                 elementType: 'div',
                 cellItemNgStyle: { 'width': '120px' },